Support full timestamp on concurrent mode failure line, for JDK8
Define variable for gc_timestamp regular expression, the expression
is without timezone, should perhaps be with timezone.

Add support for having gc_timestamp_re on concurrent mode failure
line. But the whole logic there is a bit strange, because the
time recorded for gen1t in that case is not printed at all.