Add utility for adding datestamp for gc.log files without it
Sometimes you have gc.log files without datestamps as first column,
either because you are running an old JDK (1.4 or 5), which does
not support -XX:+PrintGCDateStamps.
Or simply because you have not used -XX:+PrintGCDateStamps when
starting the java process.

This utility allows you to add the datestamp column, if you know
the start time of the java process that has generated the
gc.log files.

If you do not know the process start time, you can just pick a
date, but then you cannot correlate the gc graphs with other
graphs, for example sar graphs.